pyproject.toml
[project]
name = "airflow-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.6",
    "dbt-core>=1.11.7",
    "dbt-snowflake>=1.11.3",
    "pendulum>=3.1.0",
]
